services/arc-scarlett-voice/requirements.txt
# ==============================================================================
# arc-scarlett-voice Dependencies
# LiveKit Voice Agent with Whisper STT, Sherlock LLM (NATS), and Piper TTS
# ==============================================================================

# ------------------------------------------------------------------------------
# LiveKit Agents SDK
# ------------------------------------------------------------------------------
livekit-agents==0.8.0
livekit==0.11.0
livekit-api==0.5.0

# ------------------------------------------------------------------------------
# Speech-to-Text (Whisper)
# ------------------------------------------------------------------------------
livekit-plugins-whisper==0.6.0
faster-whisper==0.10.0  # Backend for Whisper plugin
openai-whisper==20231117  # Fallback if needed

# ------------------------------------------------------------------------------
# Voice Activity Detection (Silero)
# ------------------------------------------------------------------------------
livekit-plugins-silero==0.6.0

# ------------------------------------------------------------------------------
# Text-to-Speech (Piper)
# ------------------------------------------------------------------------------
piper-tts==1.2.0
onnxruntime==1.16.3  # Required for Piper ONNX inference

# ------------------------------------------------------------------------------
# Messaging (NATS)
# ------------------------------------------------------------------------------
nats-py==2.6.0

# ------------------------------------------------------------------------------
# Observability (OpenTelemetry)
# ------------------------------------------------------------------------------
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-exporter-otlp==1.22.0
opentelemetry-exporter-jaeger==1.22.0

# ------------------------------------------------------------------------------
# Utilities
# ------------------------------------------------------------------------------
python-dotenv==1.0.0
structlog==24.1.0  # Structured logging
tenacity==8.2.3  # Retry logic for NATS
aiofiles==23.2.1  # Async file operations
